trying to get new unittest working with CMake
