+++ Test++/Test++.vcxproj
@@ -49,7 +49,6 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="SetPluginPath.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\modules\c++\nitf-c++.vcxproj">
