+++ modules/c++/nitf-c++.vcxproj
@@ -76,6 +76,7 @@
     <ClCompile Include="nitf\source\TextSegment.cpp" />
     <ClCompile Include="nitf\source\TextSubheader.cpp" />
     <ClCompile Include="nitf\source\TRE.cpp" />
+    <ClCompile Include="nitf\source\UnitTests.cpp" />
     <ClCompile Include="nitf\source\Utils.cpp" />
     <ClCompile Include="nitf\source\WriteHandler.cpp" />
     <ClCompile Include="nitf\source\Writer.cpp" />
@@ -161,6 +162,7 @@
     <ClInclude Include="nitf\include\nitf\TextSubheader.hpp" />
     <ClInclude Include="nitf\include\nitf\TRE.hpp" />
     <ClInclude Include="nitf\include\nitf\TREField.hpp" />
+    <ClInclude Include="nitf\include\nitf\UnitTests.hpp" />
     <ClInclude Include="nitf\include\nitf\Utils.hpp" />
     <ClInclude Include="nitf\include\nitf\Version.hpp" />
     <ClInclude Include="nitf\include\nitf\WriteHandler.hpp" />
