+++ modules/c++/nitf-c++.vcxproj.filters
@@ -214,6 +214,9 @@
     <ClCompile Include="nitf\source\J2KCompressor.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="nitf\source\UnitTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
@@ -456,6 +459,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="nitf\include\nitf\J2KCompressor.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="nitf\include\nitf\UnitTests.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
