+++ modules/c/j2k/CMakeLists.txt
@@ -38,4 +38,16 @@ coda_add_tests(
         test_j2k_read_region.c
         test_j2k_read_tile.c)
 
+# Build all J2KCompress and J2KDecompress
+set(j2k_shared_srcs  J2KCompress J2KDecompress)
+foreach(j2k_shared ${j2k_shared_srcs})
+    add_library(${j2k_shared} SHARED shared/${j2k_shared}.c)
+    target_link_libraries(${j2k_shared} PUBLIC nitf-c)
+    target_compile_definitions(${j2k_shared} PRIVATE NITF_MODULE_EXPORTS)
 
+    # This line is making sure the resultant TRE is named, e.g.
+    # XML_DATA_CONTENT.so instead of libXML_DATA_CONTENT.so
+    set_target_properties(${j2k_shared} PROPERTIES PREFIX "")
+endforeach()
+
+install(TARGETS ${j2k_shared_srcs} DESTINATION "share/nitf/plugins")
